Add unit tests for all View components
Finally was able to figure out dummy router params. The view tests make sure
that the view is able to render, which would fail on a template error or the like.
Further down can explore the different states of the view, though that is probably
best left to full integration tests.