Allow DataControlService to handle week cache initialization
The week cache init from the backend is designed so the backend could provide the week
starts directly. Otherwise, for updates it isn't possible to determine the week starts. This moves
the week starts into the DataControlService so that it can generate the week starts using all
games and not just the updated data.

The HTTP backend service now follows this new model, returning `null` from  `getWeekStarts`.